server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
argon2-cffi==18.1.0       # via django
asgi-redis==1.4.3
asgiref==1.1.2            # via asgi-redis, channels, daphne
attrs==17.4.0             # via automat
autobahn==18.3.1          # via daphne
automat==0.6.0            # via twisted
babel==2.5.3
bleach-whitelist==0.0.8
bleach==2.1.3
cchardet==2.1.1           # via talon
certifi==2018.1.18        # via requests
cffi==1.11.5              # via argon2-cffi
channels==1.1.8
chardet==3.0.4            # via requests, talon
constantly==15.1.0        # via twisted
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
cssselect==1.0.3          # via talon
daphne==1.4.2             # via channels
django-anymail==2.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-enumfield==1.3b2
django-extensions==2.0.6
django-filter==1.1.0
django-influxdb-metrics==1.3.3
django-jinja==2.4.1
django-redis==4.9.0
django-rest-swagger==2.1.2
django-timezone-field==2.1
django-versatileimagefield==1.9
django[argon2]==2.0.3
djangorestframework==3.7.7
drf-nested-routers==0.90.2
factory-boy==2.10.0
faker==0.8.12             # via factory-boy
flake8==3.5.0
furl==1.0.1
gevent==1.2.2
greenlet==0.4.13          # via gevent
hiredis==0.2.0
html2text==2018.1.9
html5lib==1.0.1           # via bleach, talon
huey==1.9.0
hyperlink==18.0.0         # via twisted
idna==2.6                 # via hyperlink, requests
incremental==17.5.0       # via twisted
influxdb==5.0.0           # via django-influxdb-metrics
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema, django-jinja
lxml==4.2.1               # via talon
markdown==2.6.11
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
msgpack-python==0.5.6     # via asgi-redis
numpy==1.14.2             # via scipy, talon
openapi-codec==1.3.2      # via django-rest-swagger
orderedmultidict==0.7.11  # via furl
pillow==5.0.0             # via django-versatileimagefield
psutil==5.4.3             # via python-server-metrics
psycopg2-binary==2.7.4
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pyfcm==1.4.5
pyflakes==1.6.0           # via flake8
pymdown-extensions==4.9.2
python-dateutil==2.7.2    # via faker, influxdb
python-server-metrics==0.2.1  # via django-influxdb-metrics
pytz==2018.3
raven==6.6.0
redis==2.10.6
regex==2018.2.21          # via talon
requests-toolbelt==0.8.0  # via pyfcm
requests==2.18.4          # via coreapi, django-anymail, influxdb, pyfcm, requests-toolbelt, sparkpost
scikit-learn==0.19.1      # via talon
scipy==1.0.1              # via talon
simplejson==3.13.2        # via django-rest-swagger
six==1.11.0               # via argon2-cffi, asgi-redis, asgiref, autobahn, automat, bleach, django-anymail, django-extensions, faker, furl, html5lib, influxdb, orderedmultidict, python-dateutil, talon, tld, txaio
sparkpost==1.3.6
talon==1.4.4
text-unidecode==1.2       # via faker
tld==0.7.9                # via django-influxdb-metrics
twisted==17.9.0           # via daphne
txaio==2.9.0              # via autobahn
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
webencodings==0.5.1       # via html5lib
zope.interface==4.4.3     # via twisted
